Add DataEntityFactory and related classes (#11)
* Add DataEntityFactory and related classes

New classes for data entity factories were implemented, including DataEntityFactory, CreateFactoryData and FactoryData. Also, the `data()` method in MockResponse was refactored to handle instances of the DataEntityFactory and 'fakerphp/faker' was added to the composer.json file. A new error ignore entry was added in phpstan-baseline.neon to manage the unsafe usage of 'new static'.

* Add Data Entity Factory creation command

The commit introduces a new console command to generate Data Entity Factory classes. This is supported by the addition of a new template stub for these factory classes. The new command was also added to the command list in the DataEntitiesServiceProvider.

* Namespace