Changes to doc blocks (#26)
* Fixed docblock for data in response class

* Add methods to get response data as object and collection

Implemented two new functions in the `Response` class to return the data as an object or a collection. Also added corresponding unit tests for these functions. These enhancements provide more flexibility in handling the response data.

* Update README with extended Table of Contents and Response methods

The README.md file has been updated with an expanded Table of Contents, including new subsections on 'Casts' and more detailed 'Response methods'. The 'Response methods' section also includes new methods such as 'data with a key', 'data with a key and default value', 'as object' and 'as collection'.