Refactor MockResponse and add output handling
The code has been refactored to deal with the 'output' component in the MockResponse class. Previously, the data() function directly merged getData() and output. This has been updated to account for an empty output. A new function getOutput() has been added to deal with the output within the DataEntityFactory. Similar handling of output has been introduced in the FactoryData class too.