Update MockResponse handling and add HasMutatedData trait (#38)
* Update MockResponse handling and add HasMutatedData trait

Updated MockResponse to handle exceptions separately and added HasMutatedData trait to Response. The method in DataEntity instances has been set to SELECT by default and processing mutators has been moved from Response to AccessorProcessor. Made the class property on MockResponse and PendingQuery in Response read-only for better immutability.

* Refactor MockResponse and add output handling

The code has been refactored to deal with the 'output' component in the MockResponse class. Previously, the data() function directly merged getData() and output. This has been updated to account for an empty output. A new function getOutput() has been added to deal with the output within the DataEntityFactory. Similar handling of output has been introduced in the FactoryData class too.

* Fix