Add support for single response type (#40)
The code has been updated to handle the case where the response type is a single entity. Changes were made in 'MockProcessor.php' and 'Response.php' files to enable fetching and processing of data accordingly, when the response type is set to 'SINGLE'. Any related method calling and data handling were also adjusted for this response type.